% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adm_to_multiadm.R
\name{add_adm_to_multiadm}
\alias{add_adm_to_multiadm}
\title{add adm object ot multiadm object}
\usage{
add_adm_to_multiadm(x, ...)
}
\arguments{
\item{x}{multiadm object}

\item{...}{adm objects to be added to x}
}
\value{
a multiadm object
}
\description{
add adm object ot multiadm object
}
