% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_to_ddc.R
\name{adm_to_ddc}
\alias{adm_to_ddc}
\title{construct depth-depth curve from age-depth models}
\usage{
adm_to_ddc(adm1, adm2)
}
\arguments{
\item{adm1}{age-depth model 1. An \code{adm} object}

\item{adm2}{age-depth model 2. An \code{adm} object}
}
\description{
constructs a depth-depth curve from section 1 (described by \code{adm1}) to section 2 (described by \code{adm2})
}
