% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarboCATLite_data.R
\docType{data}
\name{CarboCATLite_data}
\alias{CarboCATLite_data}
\title{Data exported for CatrboCATLite model run}
\format{
A list with the following fields:
\itemize{
\item \emph{time_myr} : time points in Myr from the model run
\item \emph{height_2_km_offshore_m} : sediment thickness accumulated 2 km from shore
\item \emph{height_4_km_offshore_m} :  sediment thickness accumulated 4 km from shore
\item \emph{height_6_km_offshore_m} :  sediment thickness accumulated 6 km from shore
\item \emph{height_8_km_offshore_m} :  sediment thickness accumulated 8 km from shore
\item \emph{height_10_km_offshore_m} :  sediment thickness accumulated 10 km from shore
\item \emph{height_12_km_offshore_m} :  sediment thickness accumulated 2 km from shore
}
}
\source{
Add refs here
}
\usage{
CarboCATLite_data
}
\description{
Data exported for CatrboCATLite model run
}
\keyword{datasets}
