% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condensation.R
\name{condensation}
\alias{condensation}
\title{condensation with height}
\usage{
condensation(x, h, mode = "rcll", ...)
}
\arguments{
\item{x}{adm or multiadm object}

\item{h}{numeric vector, positions where condensation is determined}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
if x is an adm object, a numeric vector of condensations. if x is a multiadm object, a list of condensations
}
\description{
returns (instantaneous) condensation (time preserved per length increment) for a section
}
