% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condensation.R
\name{condensation_fun}
\alias{condensation_fun}
\title{condensation function}
\usage{
condensation_fun(x, mode = "rcll", ...)
}
\arguments{
\item{x}{adm object}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
a function
}
\description{
returns a function that determines instantaneous condensation (time preserved per strat. increment)
}
