% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_ddc.R
\name{flip_ddc}
\alias{flip_ddc}
\title{reverse direction of depth-depth curve}
\usage{
flip_ddc(x, ...)
}
\arguments{
\item{x}{a depth-depth curve (ddc object)}

\item{...}{other parameters, currently unused}
}
\description{
reverses the direction of correlation of the depth-depth curve
i.e., if the original depth-depth curve maps section 1 to section 2
the flipped curve maps section 2 to section 1
}
