% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_helpers.R
\name{flux_quad}
\alias{flux_quad}
\title{quadratic deterministix proxy flux}
\usage{
flux_quad(a = 1, b = 1, c = 1)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{c}{numeric}
}
\value{
a functional that takes no arguments. Upon each evaluation, it returns the quadratic function f(x) = ax^2 + bx + c
}
\description{
For usage with \emph{strat_cont_to_multiadm} : defines quadratic proxy flux in the time domain
defined by the function f(x) = ax^2 + bx + c
}
