% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_completeness.adm.R
\name{get_completeness.adm}
\alias{get_completeness.adm}
\title{Determine stratigraphic completeness}
\usage{
\method{get_completeness}{adm}(adm)
}
\arguments{
\item{adm}{an adm object}
}
\value{
Number between 0 and 1, the stratigraphic completeness
}
\description{
Determine stratigraphic completeness
}
