% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_completeness.R
\name{get_completeness}
\alias{get_completeness}
\title{Determine stratigraphic completeness}
\usage{
get_completeness(adm)
}
\arguments{
\item{adm}{an adm object}
}
\value{
Number between 0 and 1, the stratigraphic completeness
}
\description{
Determine stratigraphic completeness
}
