% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height.adm.R
\name{get_height.adm}
\alias{get_height.adm}
\title{determine stratigraphic height deposited at specific time}
\usage{
\method{get_height}{adm}(adm, t, destructive = TRUE, out_dom_val_h = "default", ...)
}
\arguments{
\item{adm}{adm object}

\item{t}{vector of times}

\item{destructive}{logical. if true, returns NA for times that coincide with destructive intervals}

\item{out_dom_val_h}{value assigned to values outside of the domain. If "strat_limits", outputs are limited to max/min depth of strat height}

\item{...}{parameters handed over to is_destructive}
}
\value{
a vector with same length as t, containing the strat heights deposited
}
\description{
determine stratigraphic height deposited at specific time
}
