% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height.R
\name{get_height}
\alias{get_height}
\title{determine stratigraphic height deposited at specific time}
\usage{
get_height(adm, t, destructive = TRUE)
}
\arguments{
\item{adm}{adm object}

\item{t}{vector of times}

\item{destructive}{logical. if true, returns NA for times that coincide with destructive intervals}
}
\value{
a vector with same length as t, containing the strat heights deposited
}
\description{
determine stratigraphic height deposited at specific time
}
