% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hiat_duration.R
\name{get_hiat_duration}
\alias{get_hiat_duration}
\title{extract hiatus duration}
\usage{
get_hiat_duration(x)
}
\arguments{
\item{x}{an adm object}
}
\value{
a vector with one element per hiatus: the duration of the hiatus
}
\description{
returns a vector of hiatus durations
}
\seealso{
\itemize{
\item \code{\link[=get_hiat_pos]{get_hiat_pos()}} to determine only stratigraphic position of hiatuses
\item \code{\link[=get_hiat_no]{get_hiat_no()}} to determine number of hiatuses in an adm
\item \code{\link[=get_hiat_list]{get_hiat_list()}} to get hiatus position, stat & end time
}
}
