% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_L_tp.R
\name{get_L_tp}
\alias{get_L_tp}
\title{get height/length tie point}
\usage{
get_L_tp(x, ...)
}
\arguments{
\item{x}{age-depth model (adm/multiadm) or sediment accumulation curve (sac)}

\item{...}{other options, currently not used}
}
\value{
numeric vector of the time/length tie points
}
\description{
extracts the height/length time points from an age-depth model or sediment accumulation curve
}
\seealso{
\code{\link[=get_T_tp]{get_T_tp()}} to extract time tie points
}
