% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{get_L_unit}
\alias{get_L_unit}
\title{extract length unit}
\usage{
get_L_unit(x, ...)
}
\arguments{
\item{x}{adm or multiadm object}

\item{...}{other parameters}
}
\value{
character - the length unit of \code{x}
}
\description{
extracts the length unit from sac, adm or multiadm object
}
\seealso{
\code{\link[=get_T_unit]{get_T_unit()}} \code{\link[=set_L_unit]{set_L_unit()}}
}
