% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{get_L_units}
\alias{get_L_units}
\title{extract length units}
\usage{
get_L_units(x, ...)
}
\arguments{
\item{x}{an ddc object}

\item{...}{other parameters}
}
\value{
character - the length unit of \code{x}
}
\description{
extracts the length unit from depth-depth curves (ddc objects)
}
\seealso{
\code{\link[=get_L_unit]{get_L_unit()}} \code{\link[=set_L_unit]{set_L_unit()}} to extract length units from age-depth models or sediment accumulation curves
}
