% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_names.R
\name{get_section_names}
\alias{get_section_names}
\title{get section names from depth-depth curve}
\usage{
get_section_names(x, ...)
}
\arguments{
\item{x}{a depth-depth curve (ddc object)}

\item{...}{other parameters (currently unused)}
}
\value{
vector of length 2, section names of the correlated sections
}
\description{
get section names from depth-depth curve
}
