% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_T_tp.R
\name{get_T_tp}
\alias{get_T_tp}
\title{extract time tie points}
\usage{
get_T_tp(x, ...)
}
\arguments{
\item{x}{age-depth model (adm/multiadm) or sediment accumulation curve (sac)}

\item{...}{other options, currently unused}
}
\value{
a vector, containing the time tie points
}
\description{
Extracts the time tie points from an age-depth model or sediment accumulation curve
}
\seealso{
\code{\link[=get_L_tp]{get_L_tp()}} to extract length/height tie points
}
