% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{get_T_unit}
\alias{get_T_unit}
\title{extract Time unit}
\usage{
get_T_unit(x, ...)
}
\arguments{
\item{x}{adm or multiadm object}

\item{...}{other parameters}
}
\value{
character - the time unit of \code{x}
}
\description{
extracts the Time unit from adm or multiadm object
}
\seealso{
\code{\link[=set_T_unit]{set_T_unit()}} \code{\link[=get_L_unit]{get_L_unit()}}
}
