% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.adm.R
\name{get_time.adm}
\alias{get_time.adm}
\title{determine formation time of stratigraphic positions}
\usage{
\method{get_time}{adm}(
  adm,
  h,
  hiat_mode = "start",
  bdry_pts_hiat = "destructive",
  out_dom_val_t = "default"
)
}
\arguments{
\item{adm}{an adm object}

\item{h}{vector of stratigraphic positions}

\item{hiat_mode}{string, "start", "end", or "destroy". If a stratigraphic position
coincides with a hiatus, should the start time or the end time of the hiatus be returned?
Using "destroy" retruns NA}

\item{bdry_pts_hiat}{string, "consistent" or "destructive". If the adm starts/ends with a hiatus, should
the time returned be consistent with \emph{hiat_mode}, or should it be NA?}

\item{out_dom_val_t}{:"default", "time_limits", or a numeric value. value assigned to values outside of the adms domain.
if "default", NA is used, if "time_limits",}
}
\description{
determine formation time of stratigraphic positions
}
