% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_duration.R
\name{get_total_duration}
\alias{get_total_duration}
\title{Total duration covered}
\usage{
get_total_duration(x, ...)
}
\arguments{
\item{x}{age-depth model (adm/multiadm) or sediment accumulation curve (sac)}

\item{...}{other options, currently unused}
}
\value{
numeric, total duration covered by the age-depth models/sediment accumulation curve
}
\description{
Total duration covered
}
\seealso{
\code{\link[=min_time]{min_time()}} and \code{\link[=max_time]{max_time()}} to extract the first/last tie point in time
}
