% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_thickness.R
\name{get_total_thickness}
\alias{get_total_thickness}
\title{get thickness}
\usage{
get_total_thickness(x)
}
\arguments{
\item{x}{an adm object}
}
\value{
numeric vector containing total sediment thickness accumulated
}
\description{
get thickness
}
\seealso{
\code{\link[=max_height]{max_height()}} and \code{\link[=min_height]{min_height()}} to extract the highest/lowest stratigraphic point
}
