% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_thicknesses.R
\name{get_total_thicknesses}
\alias{get_total_thicknesses}
\title{get thicknesses of sections from depth-depth curves}
\usage{
get_total_thicknesses(x, ...)
}
\arguments{
\item{x}{a depth-depth curve (ddc object)}

\item{...}{other parameters (currently unused)}
}
\value{
a numeric vector of length 2 - the thicknesses of section 1 and 2 in the depth-depth curve
}
\description{
get thicknesses of sections from depth-depth curves
}
\seealso{
\itemize{
\item \code{\link[=get_total_thickness]{get_total_thickness()}} to determine the total thickness accumulated by age-depth models or sediment accumulation curves.
}
}
