% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adm.R
\name{is_adm}
\alias{is_adm}
\title{Is an adm object a valid age-depth model}
\usage{
is_adm(x, quietly = TRUE)
}
\arguments{
\item{quietly}{logical. should descriptive warnings be shown?}

\item{adm}{an object}
}
\value{
logical. Is the input a valid adm object?
}
\description{
Contructors for adm objects such as \emph{tp_2_adm} do not check whether the inputs define a valid
age-depth mode, e.g. one where the law of superposition holds. This function performs these checks
}
