% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_multiadm.R
\name{is_multiadm}
\alias{is_multiadm}
\title{is valid multiadm object?}
\usage{
is_multiadm(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Boolean. is the object a valid multiadm object?
}
\description{
is valid multiadm object?
}
