% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sac.R
\name{is_sac}
\alias{is_sac}
\title{is valid sac objects}
\usage{
is_sac(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
logical. Is x a valid \code{sac} object?
}
\description{
checks if the object is a valid sac object
}
