% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_heigth.R
\name{max_height}
\alias{max_height}
\title{get highest stratigraphic position in an adm}
\usage{
max_height(x)
}
\arguments{
\item{x}{an age-depth model}
}
\value{
a scalar
}
\description{
get highest stratigraphic position in an adm
}
\seealso{
\code{\link[=min_height]{min_height()}}, \code{\link[=get_total_thickness]{get_total_thickness()}}
}
