% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_time.R
\name{max_time}
\alias{max_time}
\title{extract earliest time from adm}
\usage{
max_time(x)
}
\arguments{
\item{x}{age-depth model}
}
\value{
a scalar, timing of earliest tie point in the adm
}
\description{
extract earliest time from adm
}
\seealso{
\code{\link[=min_time]{min_time()}}, \code{\link[=get_total_duration]{get_total_duration()}}
}
