% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_summary.R
\name{mean_adm}
\alias{mean_adm}
\title{get mean ADM}
\usage{
mean_adm(x, h)
}
\arguments{
\item{x}{a multiadm object}

\item{h}{the heights at which to evaluate the adm}
}
\value{
an adm object
}
\description{
returns the mean adm of a multiadm object
}
