% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_summary.R
\name{median_adm}
\alias{median_adm}
\title{get median ADM}
\usage{
median_adm(x, h)
}
\arguments{
\item{x}{a multiadm object}

\item{h}{the heights at which to evaluate the adm}
}
\value{
an adm object
}
\description{
returns the median adm of a multiadm object
}
\seealso{
\code{\link[=mean_adm]{mean_adm()}} for the mean age-depth model, \code{\link[=quantile_adm]{quantile_adm()}} for the more general implementation
}
