% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_adm_to_multiadm.R
\name{merge_adm_to_multiadm}
\alias{merge_adm_to_multiadm}
\title{combine multiple adm ojects into multiadm object}
\usage{
merge_adm_to_multiadm(...)
}
\arguments{
\item{...}{adm objects}
}
\value{
object of class multiadm
}
\description{
combine multiple adm ojects into multiadm object
}
