% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_multiadm.R
\name{merge_multiadm}
\alias{merge_multiadm}
\title{merge multiple multiadm objects}
\usage{
merge_multiadm(...)
}
\arguments{
\item{...}{adm objects}
}
\value{
multiadm object
}
\description{
merge multiple multiadm objects
}
