% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_height.R
\name{min_height}
\alias{min_height}
\title{get lowest stratigraphic position in an adm}
\usage{
min_height(x)
}
\arguments{
\item{x}{an age-depth model}
}
\description{
get lowest stratigraphic position in an adm
}
\seealso{
\code{\link[=get_total_thickness]{get_total_thickness()}}, \code{\link[=max_height]{max_height()}}
}
