% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_time.R
\name{min_time}
\alias{min_time}
\title{first time tie point}
\usage{
min_time(x, ...)
}
\arguments{
\item{x}{age-depth model (adm) or sediment accumulation curve (sac)}

\item{...}{other variables, ignored}
}
\value{
number, timing of first tie point of the age-depth model/sediment accumulation curve
}
\description{
first time tie point
}
\seealso{
\code{\link[=max_time]{max_time()}}, \code{\link[=get_total_duration]{get_total_duration()}}
}
