% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adm.R
\name{plot.adm}
\alias{plot.adm}
\title{plotting adm objects}
\usage{
\method{plot}{adm}(
  x,
  lwd_hiat = 1,
  lwd_cons = 1,
  lty_hiat = 3,
  lty_cons = 1,
  col_hiat = "black",
  col_cons = "black",
  ...
)
}
\arguments{
\item{x}{an adm object}

\item{lwd_hiat}{line width of hiatuse}

\item{lwd_cons}{line width of conservative intervals}

\item{lty_hiat}{linetype of hiatuses}

\item{lty_cons}{line type of conservative intervals}

\item{col_hiat}{color of erosive intervals}

\item{col_cons}{color of conservative intervals}
}
\description{
plotting adm objects
}
