% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ddc.R
\name{plot.ddc}
\alias{plot.ddc}
\title{plot depth-depth curve}
\usage{
\method{plot}{ddc}(x, ...)
}
\arguments{
\item{x}{depth-depth curve}

\item{...}{other parameters passed to plot}
}
\description{
plot depth-depth curve
}
