% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiadm.R
\name{plot.multiadm}
\alias{plot.multiadm}
\title{plot multiadm object}
\usage{
\method{plot}{multiadm}(x, ...)
}
\arguments{
\item{x}{multiadm object}

\item{...}{parameters passed to \code{plot}}
}
\value{
a plot of the multiadm object
}
\description{
plots the median age (red) and the 95 \% envelope (blue) of a multiadm object
}
\examples{
\dontrun{
# see
vignette("adm_from_trace_cont")
# and 
vignette("adm_from_sedrate")
# for example plots.
}

}
