% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sac.R
\name{plot.sac}
\alias{plot.sac}
\title{plot sediment accumulation curve}
\usage{
\method{plot}{sac}(x, ...)
}
\arguments{
\item{x}{object of class \emph{sac}}

\item{...}{further parameters (currently ignored)}
}
\value{
invisible NULL
}
\description{
plot sediment accumulation curve
}
