% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stratlist.R
\name{plot.stratlist}
\alias{plot.stratlist}
\title{plot strat list}
\usage{
\method{plot}{stratlist}(x, orientation = "du", ord_name = "y", ...)
}
\arguments{
\item{x}{stratlist object}

\item{orientation}{character, either "du" (down-up) or "lr" (left-right). Orientation of plotting}

\item{ord_name}{name of the ordinate. Values plotted against time}

\item{...}{further arguments passed to plot}
}
\description{
plots a \code{stratlist}, i.e. a list of values associated with stratigraphic positions (typically returned by time_to_strat). will plot the element with matching \code{ord_name} against stratigraphic positions.
}
