% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.timelist.R
\name{plot.timelist}
\alias{plot.timelist}
\title{plot time lists}
\usage{
\method{plot}{timelist}(x, ...)
}
\arguments{
\item{x}{a time list}

\item{...}{other options passed to plot}
}
\description{
plot time lists
}
