% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condensation.R
\name{plot_condensation}
\alias{plot_condensation}
\title{plot condensation in height}
\usage{
plot_condensation(x, h = "default", mode = "rcll", ...)
}
\arguments{
\item{x}{an adm object}

\item{h}{"default" or a numeric vector of height where the sed rate is evaluated}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
invisible null
}
\description{
plots condensation (time per stratigraphic increment) throughout the section
}
