% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erosive_intervals.R
\name{plot_erosive_intervals}
\alias{plot_erosive_intervals}
\title{mark erosive time intervals}
\usage{
plot_erosive_intervals(
  density = NULL,
  angle = 45,
  col = "azure3",
  border = NA,
  lty = 1,
  lwd = 1
)
}
\arguments{
\item{density}{parameter passed to \emph{rect}, see ?rect for details}

\item{angle}{parameter passed to \emph{rect}, see ?rect for details}

\item{col}{parameter passed to \emph{rect}, see ?rect for details}

\item{border}{parameter passed to \emph{rect}, see ?rect for details}

\item{lty}{parameter passed to \emph{rect}, see ?rect for details}

\item{lwd}{parameter passed to \emph{rect}, see ?rect for details}
}
\value{
invisible NULL
}
\description{
mark erosive time intervals
}
