% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sed_rate.R
\name{plot_sed_rate_l}
\alias{plot_sed_rate_l}
\title{plot sed. rate in height}
\usage{
plot_sed_rate_l(x, h = "default", mode = "rcll", ...)
}
\arguments{
\item{x}{an adm object}

\item{h}{"default" or a numeric vector of height where the sed rate is evaluated}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
invisible null
}
\description{
plot sed. rate in height
}
