% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sed_rate.R
\name{plot_sed_rate_t}
\alias{plot_sed_rate_t}
\title{plot sedimentation rate}
\usage{
plot_sed_rate_t(x, mode = "rcll")
}
\arguments{
\item{x}{adm object}

\item{mode}{string, "rcll" or "lcrl". Should the sedimentation rate be Right Continuous with Left Limits (rcll) or Left Continuous with Right Limits (lcrl)}
}
\value{
invisible NULL
}
\description{
plot sedimentation rate
}
