% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_summary.R
\name{quantile_adm}
\alias{quantile_adm}
\title{get quantile ADM}
\usage{
quantile_adm(x, h, p)
}
\arguments{
\item{x}{a multiadm object}

\item{h}{the heights at which to evaluate the adm}

\item{p}{percentile}
}
\value{
an adm object
}
\description{
returns the quantile adm of a multiadm object
}
