% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_dir.R
\name{rev_dir}
\alias{rev_dir}
\title{reverse direction of time/depth axis}
\usage{
rev_dir(x, ref)
}
\arguments{
\item{x}{object to transform - typically a \code{fossil} or \code{taxonomy} object}

\item{ref}{reference point used for reversal}
}
\description{
The \code{FossilSim} package simulates fossils, trees, and taxonomies using age
meaning 0 represents the present and larger numbers indicate older ages
To interact with \code{admtools}, the direction of time must be reversed
effectively replaces the time component \code{t} of an object by \code{ref - t}
}
\examples{
\dontrun{
  # for usage example, see
  vignette("FossilSim_integration")
}


}
