% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_to_adm.R
\name{sac_to_adm}
\alias{sac_to_adm}
\title{turn sed. acc curve into adm}
\usage{
sac_to_adm(x)
}
\arguments{
\item{x}{object of class \emph{sac}}
}
\value{
object of class \emph{adm}
}
\description{
turn sed. acc curve into adm
}
\seealso{
\code{\link[=tp_to_adm]{tp_to_adm()}} for the generator of \emph{adm}
}
