% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedrate_gen_helpers.R
\name{sed_rate_gen_from_bounds}
\alias{sed_rate_gen_from_bounds}
\title{seg rate gen from upper/lower bounds}
\usage{
sed_rate_gen_from_bounds(h_l, s_l, h_u, s_u, rate = 1)
}
\arguments{
\item{h_l}{height values for lower bounds}

\item{s_l}{sed rate values for lower bounds}

\item{h_u}{height values for upper bounds}

\item{s_u}{sed rate values for upper bounds}

\item{rate}{rate of poisson point process}
}
\value{
a function factory for usage with \code{sedrate_to_multiadm}
}
\description{
seg rate gen from upper/lower bounds
}
\seealso{
\code{\link[=sedrate_to_multiadm]{sedrate_to_multiadm()}} for estimating age-depth models using the outputs, \code{\link[=sed_rate_from_matrix]{sed_rate_from_matrix()}} for other means of defining sedimentation rates
}
