% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_rate_h.R
\name{sed_rate_l}
\alias{sed_rate_l}
\title{sedimentation rate in stratigraphic height}
\usage{
sed_rate_l(x, h, mode = "rcll", ...)
}
\arguments{
\item{x}{adm object}

\item{h}{numeric vector, stratigraphic positions}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
a vector of sed rates (if x is an adm object), or a list of sedimentation rates
}
\description{
determines instantaneous sedimentation rate at a specified stratigraphic position
}
