% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_rate_l.R
\name{sed_rate_l_fun}
\alias{sed_rate_l_fun}
\title{sed rate in height function}
\usage{
sed_rate_l_fun(x, mode = "rcll", ...)
}
\arguments{
\item{x}{an adm object}

\item{mode}{string, handed over to \emph{sed_rate_t}, see ?sed_rate_t for details}

\item{...}{parameters passed to \emph{get_time}, see ?get_time for details}
}
\value{
a function
}
\description{
returns a function that determines sed. rates in height
}
