% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_rate_t.R
\name{sed_rate_t_fun}
\alias{sed_rate_t_fun}
\title{sedimentation rate function}
\usage{
sed_rate_t_fun(x, mode = "rcll")
}
\arguments{
\item{x}{an adm object}

\item{mode}{string, "rcll" or "lcrl". Should the sedimentation rate be Right Continuous with Left Limits (rcll) or Left Continuous with Right Limits (lcrl)}
}
\value{
a function
}
\description{
returns a function that retruns sedimentation rate
}
