% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{set_L_unit}
\alias{set_L_unit}
\title{set length units}
\usage{
set_L_unit(x, L_unit, ...)
}
\arguments{
\item{x}{adm or multiadm object}

\item{L_unit}{time unit}

\item{...}{further parameters}
}
\value{
an adm or multiadm object with the L unit assigned
}
\description{
set length units for adm and multiadm objects
}
\seealso{
\code{\link[=set_T_unit]{set_T_unit()}} \code{\link[=get_L_unit]{get_L_unit()}}
}
