% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{set_L_units}
\alias{set_L_units}
\title{set length units}
\usage{
set_L_units(x, L_units, ...)
}
\arguments{
\item{x}{depth-depth curve (ddc object)}

\item{L_units}{length units, vector of length 2. i-th element is the length unit of i-th section}

\item{...}{further parameters}
}
\value{
an ddc object with the L unit assigned
}
\description{
set length units for depth-depth curves
}
\seealso{
\code{\link[=set_L_unit]{set_L_unit()}} \code{\link[=get_L_units]{get_L_units()}}
}
