% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_names.R
\name{set_section_names}
\alias{set_section_names}
\title{set section names for depth-depth curves}
\usage{
set_section_names(x, sec_names, ...)
}
\arguments{
\item{x}{a depth-depth curve (ddc object)}

\item{sec_names}{vector of length 2, section names assigned to the depth-depth curve}

\item{...}{other parameters (currently unused)}
}
\value{
a depth-depth curve with assigned section names
}
\description{
set section names for depth-depth curves
}
