% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_L_unit_handling.R
\name{set_T_unit}
\alias{set_T_unit}
\title{set time units}
\usage{
set_T_unit(x, T_unit, ...)
}
\arguments{
\item{x}{sac, adm or multiadm object}

\item{T_unit}{time unit}

\item{...}{further parameters}
}
\value{
an sac, adm or multiadm object with the time unit assigned
}
\description{
set time units for sac, adm and multiadm objects
}
\seealso{
\code{\link[=set_L_unit]{set_L_unit()}} \code{\link[=get_T_unit]{get_T_unit()}}
}
