% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_multiadm.R
\name{split_multiadm}
\alias{split_multiadm}
\title{split multiadm objects into adm}
\usage{
split_multiadm(x)
}
\arguments{
\item{x}{a multiadm object}
}
\value{
list with objects of class adm
}
\description{
split multiadm objects into adm
}
