% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_to_time.fossils.R
\name{strat_to_time.fossils}
\alias{strat_to_time.fossils}
\title{transform \code{fossils} object from \code{FossilSim} between time and strat domain}
\usage{
\method{strat_to_time}{fossils}(obj, x, ...)
}
\arguments{
\item{obj}{the \code{fossils} object}

\item{x}{the age-depth model}

\item{...}{further parameters passed to _get_height`}
}
\description{
transforms \code{fossil} objects from stratigraphic to time domain
`
}
