% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_to_time.numeric.R
\name{strat_to_time.numeric}
\alias{strat_to_time.numeric}
\title{transform numeric vectors from height to time domain}
\usage{
\method{strat_to_time}{numeric}(obj, x, ...)
}
\arguments{
\item{obj}{a numeric vector representing stratigraphic positions.}

\item{x}{an \emph{adm} object}

\item{...}{options passed to \code{get_time}}
}
\value{
A numeric vector with times of deposition of the entries in \code{obj}
}
\description{
This function transforms numeric vectors from the stratigrahpic to the time domain
Fundamentally a wrapper around get_time for consistent syntax
}
\examples{
# see vignette("admtools") for an example

}
\seealso{
\code{\link[=time_to_strat.numeric]{time_to_strat.numeric()}} for the transformation from time to height domain, \code{\link[=get_time]{get_time()}} for the underlying procedure, \code{\link[=time_to_strat]{time_to_strat()}} for the higher level function, \code{\link[=strat_to_time.list]{strat_to_time.list()}} and \code{\link[=strat_to_time.phylo]{strat_to_time.phylo()}} for the transformation of lists and phylogenetic trees.
}
