% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_to_time.R
\name{strat_to_time}
\alias{strat_to_time}
\title{transform objects from strat. to time domain}
\usage{
strat_to_time(obj, x, ...)
}
\arguments{
\item{obj}{the object to be transformed}

\item{x}{age-depth model}

\item{...}{other parameters}
}
\value{
an object of the same type as obj
}
\description{
Takes an object and transforms it from the time domain into the stratigraphic domain using the provided age-depth model.
Currently implemented for the "phylo" class.
}
