% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_to_time.taxonomy.R
\name{strat_to_time.taxonomy}
\alias{strat_to_time.taxonomy}
\title{transform taxonomy object from strat to time domain}
\usage{
\method{strat_to_time}{taxonomy}(obj, x, ...)
}
\arguments{
\item{obj}{\code{taxonomy} object as defined by the \code{FossilSim} package}

\item{x}{age-depth model}

\item{...}{further parameters passed to \emph{get_time}}
}
\description{
transform taxonomy object from strat to time domain
}
