% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adm.R
\name{summary.adm}
\alias{summary.adm}
\title{summary of age-depth model}
\usage{
\method{summary}{adm}(object, ...)
}
\arguments{
\item{object}{an adm object}

\item{...}{other variables, are ignored}
}
\value{
Invisible NULL, prints summary to the console
}
\description{
Displays some summary numbers of an age-depth models
}
\examples{
my_adm = tp_to_adm(t = 1:5, h = c(2,2,3), L_unit = "m", T_unit = "Myr" )
summary(my_adm)


}
