% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ddc}
\alias{summary.ddc}
\title{summary of depth-depth curve}
\usage{
\method{summary}{ddc}(object, ...)
}
\arguments{
\item{object}{depth-depth curve (\code{ddc})}

\item{...}{other variables, currently ignored}
}
\value{
invisible NULL
}
\description{
displays some summary numbers of a depth-depth curve (\code{ddc} object)
}
