% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.sac}
\alias{summary.sac}
\title{summary of sediment accumulation curve}
\usage{
\method{summary}{sac}(object, ...)
}
\arguments{
\item{object}{sediment accumulation curve (sac)}

\item{...}{other variables, are ignored}
}
\value{
invisible NULL
}
\description{
displays some summary numbers of sediment accumulation curve
}
