% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_strat.fossils.R
\name{time_to_strat.fossils}
\alias{time_to_strat.fossils}
\title{transform \code{fossils} from FossilSim from time to strat domain}
\usage{
\method{time_to_strat}{fossils}(obj, x, ...)
}
\arguments{
\item{obj}{the fossils object}

\item{x}{the age-depth model to be used, and \emph{adm} object}

\item{...}{further parameters to be passed to \code{to_height}, e.g., \code{is_destructive}}
}
\value{
a \code{fossils} object for usage with the \code{FossilSim} package
}
\description{
The \code{fossils} object of the \code{FossilSim} package describe the location of fossil occurrences along a phylogeny. This function transforms the objects from the time to the stratigraphic domain.
}
