% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_strat.taxonomy.R
\name{time_to_strat.taxonomy}
\alias{time_to_strat.taxonomy}
\title{transform \code{taxomony} from time to strat domain}
\usage{
\method{time_to_strat}{taxonomy}(obj, x, ...)
}
\arguments{
\item{obj}{a \code{taxonomy} object used by \code{FossilSim}}

\item{x}{an age-depth model object}

\item{...}{further parameters passed to \code{get_height}, e.g., \emph{is_destructive}}
}
\description{
transforms \code{taxonomy} objects from time to strat domain. Ignored the \code{destructive} argument of \emph{get_height}, which is automatically set to FALSE
}
