% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetree.R
\docType{data}
\name{timetree}
\alias{timetree}
\title{example time tree}
\format{
An object of class \code{phylo} of length 4.
}
\usage{
timetree
}
\description{
Time tree generated using the \code{ape} package. Code used to generate is

set.seed(1)
tree_in_time = ape::rlineage(birth = 1.8,
death = 0.2,
Tmax = 2)
}
\keyword{datasets}
