% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_2_adm.R
\name{tp_2_adm}
\alias{tp_2_adm}
\title{Turn tie points into age-depth model}
\usage{
tp_2_adm(t, h)
}
\arguments{
\item{t}{tie points in time}

\item{h}{tie points in height}

\item{T_unit}{Time unit}

\item{L_unit}{Length unit}
}
\description{
Turn tie points into age-depth model
}
