% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_helpers.R
\name{tp_time_det}
\alias{tp_time_det}
\title{deterministic tie points in time domain}
\usage{
tp_time_det(times)
}
\arguments{
\item{times}{numeric vector, times of the tie points}
}
\value{
a function for usage with \emph{strat_cont_to_multiadm} and \emph{sedrate_to_mulitadm} as t_tp input
}
\description{
defines deterministc tie points in time.
}
\seealso{
\code{\link[=tp_height_det]{tp_height_det()}} for deterministic tie points in height, \code{\link[=tp_time_norm]{tp_time_norm()}} for tie points following a normal distribution
}
