% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_helpers.R
\name{tp_time_floating_scale}
\alias{tp_time_floating_scale}
\title{tie points for floating time scale}
\usage{
tp_time_floating_scale()
}
\value{
a functional
}
\description{
Defines tie points for a floating time sclae for usage with \emph{sedrate_to_multiadm} and \emph{strat_cont_to_multiadm}
}
