% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_helpers.R
\name{tp_time_norm}
\alias{tp_time_norm}
\title{time tie points with normal distribution}
\usage{
tp_time_norm(mean, sd, force_order = TRUE)
}
\arguments{
\item{mean}{numeric vector, mean age of tie points}

\item{sd}{numeric vector, standard deviation of tie points}

\item{force_order}{logical, enforce strictly increasing times}
}
\value{
a function
}
\description{
defines a function factory that returns normally distributed times. FOr usage with \code{sedrate_to_multiadm} and \code{strat_cont_to_multiadm}.
}
