% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_to_ddc.R
\name{tp_to_ddc}
\alias{tp_to_ddc}
\title{construct depth-depth curve from coeval heights}
\usage{
tp_to_ddc(
  h1,
  h2,
  L_unit_1 = NULL,
  L_unit_2 = NULL,
  sec_1 = "unnamed section 1",
  sec_2 = "unnamed section 2"
)
}
\arguments{
\item{h1}{height tie points in section 1}

\item{h2}{height tie points in section 2}

\item{L_unit_1}{Length unit in section 1}

\item{L_unit_2}{Length unit in section 2}

\item{sec_1}{name of section 1}

\item{sec_2}{name of section 2}
}
\description{
constructs a depth-depth curve (\code{ddc} object) from a vector of coeval heights specified by \code{h1} and \code{h2}
}
