% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admtools-package.R
\docType{package}
\name{admtools-package}
\alias{admtools}
\alias{admtools-package}
\title{admtools: Estimate and Manipulate Age-Depth Models}
\description{
Estimate age-depth models from stratigraphic and sedimentological data, and transform data between the time and stratigraphic domain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MindTheGap-ERC/admtools}
  \item \url{https://mindthegap-erc.github.io/admtools/}
  \item Report bugs at \url{https://github.com/MindTheGap-ERC/admtools/issues}
}

}
\author{
\strong{Maintainer}: Emilia Jarochowska \email{e.jarochowska@uni-muenster.de} (\href{https://orcid.org/0000-0001-8937-9405}{ORCID})

Authors:
\itemize{
  \item Niklas Hohmann \email{N.H.Hohmann@uu.nl} (\href{https://orcid.org/0000-0003-1559-1838}{ORCID})
}

}
\keyword{internal}
