% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_helpers.R
\name{flux_const}
\alias{flux_const}
\title{constant deterministic proxy flux}
\usage{
flux_const()
}
\value{
a functional
}
\description{
For usage with \emph{strat_cont_to_multiadm}; defines constant proxy flux in the time domain
}
