% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_helpers.R
\name{flux_linear}
\alias{flux_linear}
\title{linear deterministic proxy flux}
\usage{
flux_linear(x0 = 0, y0 = 1, x1 = 1, y1 = 2)
}
\arguments{
\item{x0}{numeric, abscissa}

\item{y0}{numeric, ordinate}

\item{x1}{numeric, abscissa}

\item{y1}{numeric, ordinate}
}
\value{
a functional that takes no arguments. Upon each evaluation, it returns a linear function passing thrpugh the points (x0, y0) and (x1, y1)
}
\description{
For usage with \emph{strat_cont_to_multiadm} : defines linear proxy flux in the time domain
Proxy flux is the linear function passing thrpugh the points (x0, y0) and (x1, y1)
}
