% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTimeOpt_functionality.R
\name{get_data_from_eTimeOpt}
\alias{get_data_from_eTimeOpt}
\title{extract data from eTimeOpt results}
\usage{
get_data_from_eTimeOpt(res, index = 1)
}
\arguments{
\item{res}{results generated by eTimeOpt}

\item{index}{which output should be extracted? See description for details}
}
\value{
a list with three entries
"sed_rate" : numeric vector, sedimentation rates
"height" : numeric vector, heights
"results" : matrix with length(height) rows and length(sed_rate) columns. results of eTimeOpt
}
\description{
Extracts data from eTimeOpt. The type of data extracted depends on the \code{output} setting used for eTimeOpt. If you want ot extract specific data, adjust the output parameter in eTimeOpt to return the correct data (e.g. 2 for r^2 envelope). See eTimeOpt documentation for details on this. Then call this function on the return variable.
}
\seealso{
\code{\link[=sed_rate_from_matrix]{sed_rate_from_matrix()}} to use define sedimentation rates based on this functions outputs, \code{\link[=sedrate_to_multiadm]{sedrate_to_multiadm()}} to estimate age-depth models from the outputs.
}
