% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adm.R
\name{is_adm}
\alias{is_adm}
\title{Checks for validity of adm object}
\usage{
is_adm(adm)
}
\arguments{
\item{adm}{an object}
}
\value{
logical. Is the input a valid adm object?
}
\description{
Checks for validity of adm object
}
