% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adm.R
\name{is_adm}
\alias{is_adm}
\title{Is an adm object a valid age-depth model}
\usage{
is_adm(x, quietly = TRUE)
}
\arguments{
\item{x}{an object}

\item{quietly}{logical. should descriptive warnings be shown?}
}
\value{
logical. Is the input a valid adm object?
}
\description{
Constructors for adm objects such as \emph{tp_to_adm} do not check whether the inputs define a valid
age-depth mode, e.g. one where the law of superposition holds. This function performs these checks
}
\examples{

x = tp_to_adm(t = c(2,1), h = c(1,2)) # reversed order of time tie points
is_adm(x) # returns FALSE


}
