% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_destructive.R
\name{is_destructive}
\alias{is_destructive}
\title{Is deposition destructive?}
\usage{
is_destructive(adm, t, mode = "rcll")
}
\arguments{
\item{adm}{an adm object}

\item{t}{vector of times}

\item{mode}{string, either "rcll", "lcrl", "open", or "closed"}
}
\value{
logical vector of same length as t. Is deposition at time t destructive?
}
\description{
Is deposition destructive?
}
