% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_heigth.R
\name{max_height}
\alias{max_height}
\title{get highest stratigraphic tie point}
\usage{
max_height(x, ...)
}
\arguments{
\item{x}{age-depth model (adm) or sediment accumulation curve (sac)}

\item{...}{other variables, ignored}
}
\value{
number, stratigraphic position of the highest stratigraphic tie point
}
\description{
get highest stratigraphic tie point
}
\seealso{
\code{\link[=min_height]{min_height()}}, \code{\link[=get_total_thickness]{get_total_thickness()}}
}
