% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_time.R
\name{max_time}
\alias{max_time}
\title{last time tie point}
\usage{
max_time(x, ...)
}
\arguments{
\item{x}{age-depth model (adm) or sediment accumulation curve (sac)}

\item{...}{other variables, ignored}
}
\value{
number, last time tie point of the age-depth model/sediment accumulation curve
}
\description{
last time tie point
}
\seealso{
\code{\link[=min_time]{min_time()}}, \code{\link[=get_total_duration]{get_total_duration()}}
}
