% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_time.R
\name{min_time}
\alias{min_time}
\title{extract last time from adm}
\usage{
min_time(x)
}
\arguments{
\item{x}{age-depth model}
}
\value{
a scalar, timing of last tie point in the adm
}
\description{
extract last time from adm
}
\seealso{
\code{\link[=max_time]{max_time()}}, \code{\link[=get_total_duration]{get_total_duration()}}
}
