% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adm.R
\name{plot.adm}
\alias{plot.adm}
\title{plotting adm objects}
\usage{
\method{plot}{adm}(
  x,
  lwd_destr = 1,
  lwd_acc = 1,
  lty_destr = 3,
  lty_acc = 1,
  col_destr = "black",
  col_acc = "black",
  ...
)
}
\arguments{
\item{x}{an adm object}

\item{lwd_destr}{line width of hiatuse}

\item{lwd_acc}{line width of conservative intervals}

\item{lty_destr}{linetype of hiatuses}

\item{lty_acc}{line type of conservative intervals}

\item{col_destr}{color of erosive intervals}

\item{col_acc}{color of conservative intervals}

\item{...}{arguments passed to plot}
}
\description{
plotting adm objects
}
\seealso{
\code{\link[=L_axis_lab]{L_axis_lab()}} and \code{\link[=T_axis_lab]{T_axis_lab()}} for plotting time and axis labels, the vignette on plotting available via \code{browseVignettes(package = "admtools")}
}
