% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiadm.R
\name{plot.multiadm}
\alias{plot.multiadm}
\title{plot multiadm object}
\usage{
\method{plot}{multiadm}(x, ...)
}
\arguments{
\item{x}{multiadm object}

\item{...}{parameters passed to \code{plot}}
}
\value{
a plot of the multiadm object
}
\description{
plot multiadm object
}
