% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedrate_to_multiadm.R
\name{sedrate_to_multiadm}
\alias{sedrate_to_multiadm}
\title{Estimate age-depth model from sedimentation rate & tie points}
\usage{
sedrate_to_multiadm(
  h_tp,
  t_tp,
  sed_rate_gen,
  h,
  no_of_rep = 100L,
  subdivisions = 100L,
  T_unit = NULL,
  L_unit = NULL
)
}
\arguments{
\item{h_tp}{: function, returns stratigraphic positions of tie points}

\item{t_tp}{: function, returns times of deposition of tie points}

\item{sed_rate_gen}{: function, returns sedimentation rate functions}

\item{h}{: numeric, heights where the adm is calculated}

\item{no_of_rep}{: numeric, number of repetitions}

\item{subdivisions}{maximum no of subintervals used in numeric integration}

\item{T_unit}{time unit}

\item{L_unit}{length unit}
}
\value{
object of class multiadm
}
\description{
Estimate age-depth model from sedimentation rate & tie points
}
\examples{
\dontrun{
# see this vignette for an example
vignette("adm_from_sedrate")
}

}
