% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_multiadm.R
\name{split_multiadm}
\alias{split_multiadm}
\title{split multiadm objects into adm}
\usage{
split_multiadm(multiadm)
}
\arguments{
\item{multiadm}{a multiadm object}
}
\value{
a list of adm objects
}
\description{
split multiadm objects into adm
}
