% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multiadm.R
\name{summary.multiadm}
\alias{summary.multiadm}
\title{summary of age-depth model}
\usage{
\method{summary}{multiadm}(object, ...)
}
\arguments{
\item{object}{a multiadm object}

\item{...}{other variables, are ignored}
}
\value{
Prints summary to the console and returns an invisible NULL
}
\description{
Displays some summary numbers of an age-depth models
}
