% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_strat.R
\name{time_to_strat}
\alias{time_to_strat}
\title{transform objects from time domain to strat. domain}
\usage{
time_to_strat(obj, x, ...)
}
\arguments{
\item{obj}{the object to be transformed}

\item{x}{age-depth model for the transformation}

\item{...}{other parameters}
}
\value{
an object of the same type as obj
}
\description{
Takes an object and transforms it from the time domain into the stratigraphic domain using the provided age-depth model.
Currently implemented for the "phylo", "list", and "numeric" class.
}
\seealso{
\code{\link[=strat_to_time]{strat_to_time()}} to transform data from the stratigraphic domain to the time domain, \code{\link[=time_to_strat.phylo]{time_to_strat.phylo()}}, \code{\link[=time_to_strat.numeric]{time_to_strat.numeric()}} and \code{\link[=time_to_strat.list]{time_to_strat.list()}} for details on how to transfrom phylo objects, vectors, and lists. See \code{\link[=get_height]{get_height()}} for the underlying procedure.
}
