% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_helpers.R
\name{tp_height_det}
\alias{tp_height_det}
\title{deterministic tie points height domain}
\usage{
tp_height_det(heights)
}
\arguments{
\item{heights}{numeric vector. Stratigraphic positions of the tie points}
}
\value{
a function for usage with \emph{strat_cont_to_multiadm} and \emph{sedrate_to_multiamd}
}
\description{
defines deterministic stratigraphic tie points
}
