% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_helpers.R
\name{tp_time_floating_scale}
\alias{tp_time_floating_scale}
\title{tie points for floating time scale}
\usage{
tp_time_floating_scale()
}
\value{
function for usage with \emph{strat_cont_to_multiadm} and \emph{sedrate_to_multiamd} as t_tp input
}
\description{
Defines tie points for a floating (auxiliary) time scale for usage with \emph{sedrate_to_multiadm} and \emph{strat_cont_to_multiadm} as t_tp input. This floating time scale consists of two tie points in time, the first at time t = 0, the second at time t = 1. \code{tp_time_floating_scale} is a synonym of  \code{tp_time_det(times = c(0,1))}
}
\examples{
\dontrun{
# see this vignette for an example
vignette("adm_from_trace_cont")
} 
}
\seealso{
\code{\link[=tp_time_norm]{tp_time_norm()}} for tie points following a normal distribution, \code{\link[=tp_height_det]{tp_height_det()}} for deterministic height tie points
}
