% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_to_sac.R
\name{tp_to_sac}
\alias{tp_to_sac}
\title{define sed. acc. curve}
\usage{
tp_to_sac(t, h, T_unit = NULL, L_unit = NULL)
}
\arguments{
\item{t}{numeric vector, time coordinates of tie points}

\item{h}{numeric vector, height coordinates of tie points}

\item{T_unit}{time unit}

\item{L_unit}{length unit}
}
\value{
a \emph{sac} object reflecting a sediment accumulation curve
}
\description{
defines \emph{sac} (sediment accumulation curve) object from tie points
}
\seealso{
\code{\link[=sac_to_adm]{sac_to_adm()}} to transform sediment accumulation curves into age-depth models,  \code{\link[=get_T_tp]{get_T_tp()}} and \code{\link[=get_L_tp]{get_L_tp()}} to extract time and height/length tie points
}
